net/mana: fix counter overflow for posted WQE
[ upstream commit 8a56ff3d246899b77b81c6450680214a8697749b ]

This counter should set to uint16_t, the same type as pkt_received.
Otherwise, it may overflow when pkt_received goes over 256.

Thanks Xinhao Kong <t-kongxinhao@microsoft.com> for debugging this.

Fixes: 517ed6e2d590 ("net/mana: add basic driver with build environment")

+++ drivers/net/mana/rx.c
@@ -380,7 +380,7 @@ uint16_t
 mana_rx_burst(void *dpdk_rxq, struct rte_mbuf **pkts, uint16_t pkts_n)
 {
 	uint16_t pkt_received = 0;
-	uint8_t wqe_posted = 0;
+	uint16_t wqe_posted = 0;
 	struct mana_rxq *rxq = dpdk_rxq;
 	struct mana_priv *priv = rxq->priv;
 	struct rte_mbuf *mbuf;
